# docs/openapi/smartphoto_backend_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "smartphoto-backend",
    "description": "SmartPhoto Backend v2 OpenAPI。系统围绕前端 6 步流程设计，覆盖 session 创建、图片上传、分析、平台选择、copy 编辑、策略预览、Prompt 调试、整组生图、结果查询、下载、重生成。",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v2/assets/{asset_id}/regenerate": {
      "post": {
        "tags": [
          "assets"
        ],
        "summary": "单图重生成",
        "description": "基于指定 asset 触发单图重生成。当前会继承该图对应的 role/display_order，并创建新的 generation job。支持 `Idempotency-Key`，也会受 generation 并发保护。",
        "operationId": "regenerateAsset",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "可选幂等键。",
              "title": "Idempotency-Key"
            },
            "description": "可选幂等键。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetRegenerateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_GenericGenerationJobData_"
                }
              }
            }
          },
          "400": {
            "description": "业务校验失败，例如状态不合法、参数缺失、平台无效。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证。当前开发态通常不会出现。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "无权限访问该资源。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在，例如 session/job/asset 不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "并发冲突或幂等冲突，例如 40901 / 40902。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "请求体或表单参数未通过 FastAPI/Pydantic 校验。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务内部异常。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "上游 LLM 或生图服务异常，例如 50201 / 50202。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIErrorResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务错误码，如 40002 / 40901 / 50202。",
            "examples": [
              40002
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "错误描述。",
            "examples": [
              "invalid_session_status"
            ]
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "错误附加数据。当前多数场景为空。"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "APIErrorResponse"
      },
...
      "APIResponse_GenericGenerationJobData_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务状态码。成功固定为 0。",
            "default": 0,
            "examples": [
              0
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "业务消息。成功固定为 success。",
            "default": "success",
            "examples": [
              "success"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/GenericGenerationJobData",
            "description": "接口返回的业务数据。"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "APIResponse[GenericGenerationJobData]"
      },
...
      "AssetRegenerateRequest": {
        "properties": {
          "instruction": {
            "type": "string",
            "title": "Instruction",
            "description": "单图重生成指令。"
          },
          "keep_style_consistency": {
            "type": "boolean",
            "title": "Keep Style Consistency",
            "description": "是否保持与当前版本风格一致。",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "instruction"
        ],
        "title": "AssetRegenerateRequest"
      },
...
      "GenericGenerationJobData": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id",
            "description": "任务 ID。"
          },
          "job_type": {
            "type": "string",
            "title": "Job Type",
            "description": "任务类型。"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "任务状态。"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "job_type",
          "status"
        ],
        "title": "GenericGenerationJobData"
      },
...
    }
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local development"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local development alias"
    }
  ]
}